SeaweedFSClient.jl

Some of these actions will throw exceptions via the HTTP client e.g. doing delete(f) on a deleted file

SeaweedFSClient.create — Function

create(h::Host)

Arguments

h an instance of NamedTuple{(:hostname, :port), Tuple{String, Int64}}

Example

julia> f = create((hostname="localhost", port=9333))
(host = (hostname = "localhost", port = 9333), status = 200, url = "192.168.1.9:8080", publicUrl = "192.168.1.9:8080", count = 1, fid = "3,09b8e10215"

SeaweedFSClient.filedata — Function

filedata(filename, text)
filedata(filename::AbstractString)
filedata(filename::AbstractString, io::IO)

Create a file data object to send to the server

Example

# using a created f
julia> save(f, filedata("/home/matt/readme.txt"))
(host = (hostname = "localhost", port = 9333), status = 201, eTag = "04fd3fe2", name = "readme.txt", size = 29)

SeaweedFSClient.save — Function

save(f::NamedTuple, file_data)
save(h::Host, url, fid, file_data)

Arguments

f is the response returned from create (or any tuple with (host=(hostname=String, port=Int), fid="V,FID", url="http...")) Send the file_data to the fid. file_data is created using filedata()

Example

# using a created f
julia> save(f, filedata("readme.txt", "This is some text"))
(host = (hostname = "localhost", port = 9333), status = 201, eTag = "0d79b2b0", size = 17)

SeaweedFSClient.load — Function

load(f::NamedTuple, options=Dict{String,String}())::Vector{UInt8}

Example

```

using a created f

julia> load(f) 17-element Vector{UInt8}: 0x54 0x68 ⋮ 0x20 0x74 ```

SeaweedFSClient.delete — Function

delete(f::NamedTuple) 
delete(url, fid)

Arguments

f is the response returned from create (or any tuple with (host=(hostname=String, port=Int), fid="V,FID", url="http...")) Delete a file from the server

Example

# using a created f
julia> delete(f)
(host = nothing, status = 202, size = 29)